+++ includes/class-cf7-custom-messages-activator.php
@@ -19,7 +19,7 @@ class CF7_Custom_Messages_Activator {
         if ( ! is_plugin_active( 'contact-form-7/wp-contact-form-7.php' ) && ! is_plugin_active_for_network( 'contact-form-7/wp-contact-form-7.php' ) ) {
             deactivate_plugins( CF7_CUSTOM_MESSAGES_PLUGIN_BASENAME );
             wp_die(
-                esc_html__( 'This plugin requires Contact Form 7 to be installed and activated.', 'cf7-custom-messages' ),
+                esc_html__( 'This plugin requires Contact Form 7 to be installed and activated.', 'cf7-custom-validation-messages' ),
                 'Plugin dependency check',
                 array( 'back_link' => true )
             );